Move G-K alg from `sql/core` to `sql/catalyst`
